chore: move testing to jest
