+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "unionfs",
-  "version": "2.0.4",
+  "version": "2.0.6",
   "description": "Use multiple `fs` modules in a union.",
   "main": "lib/index.js",
   "keywords": [
@@ -21,56 +21,39 @@
     "build": "npm run build-ts && npm run build-js",
     "build-ts": "./node_modules/.bin/gulp build-ts",
     "build-js": "./node_modules/.bin/babel src --out-dir lib",
-    "test": "npm run test-coverage",
-    "test-basic": "./node_modules/.bin/mocha --require ts-node/register src/**/*.test.ts src/**/*.test.tsx",
-    "test-watch": "./node_modules/.bin/mocha --require ts-node/register src/**/*.test.ts src/**/*.test.tsx --watch",
-    "test-coverage": "nyc --per-file mocha --require ts-node/register --require source-map-support/register --full-trace --bail src/**/*.test.ts"
+    "test": "jest",
+    "test-watch": "jest --watch",
+    "test-coverage": "jest --coverage"
   },
   "dependencies": {
-    "fs-monkey": "^0.2.2",
-    "memfs": "^2.0.6"
+    "fs-monkey": "^0.2.2"
   },
   "devDependencies": {
-    "@types/chai": "4.0.1",
-    "@types/mocha": "2.2.41",
+    "@types/jest": "^21.1.4",
     "@types/node": "8.0.17",
     "babel-cli": "6.24.1",
     "babel-preset-es2015": "6.24.1",
-    "chai": "4.1.0",
     "gulp": "3.9.1",
     "gulp-typescript": "3.2.1",
     "memory-fs": "^0.4.1",
-    "mocha": "3.4.2",
-    "nyc": "11.1.0",
+    "memfs": "^2.5.7",
+    "jest": "^21.2.1",
     "source-map-support": "0.4.15",
-    "ts-node": "3.3.0",
-    "typescript": "2.4.2"
+    "ts-node": "^3.3.0",
+    "ts-jest": "^21.1.0",
+    "typescript": "^2.5.3"
   },
-  "nyc": {
-    "per-file": true,
-    "include": [
-      "src/**/*.ts",
-      "src/**/*.tsx"
+  "jest": {
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx"
     ],
-    "exclude": [
-      "src/**/*.test.ts",
-      "src/**/*.test.tsx"
-    ],
-    "extension": [
-      ".ts",
-      ".tsx"
-    ],
-    "require": [
-      "ts-node/register"
-    ],
-    "reporter": [
-      "text",
-      "json",
-      "lcov",
-      "text-summary"
-    ],
-    "sourceMap": true,
-    "instrument": true,
-    "cache": true
+    "transformIgnorePatterns": [],
+    "transform": {
+      "^.+\\.tsx?$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": ".*/__tests__/.*\\.(test|spec)\\.(jsx?|tsx?)$"
   }
 }
